Resolve conflict with other plugin frameworks when hooking ProcessUserCmds
Delay detouring ProcessUserCmds until after other plugin/frameworks have
had a chance to hook it, such as CS2Fixes. If some other plugin hooked it,
then use a fallback input method.

This fallback method is subpar in that:
- It does not prevent weapon switching when number keys bound
- It does not freeze the player in place when using WASD input

As it stands, just delaying hooking results in a successful hook by CS# too,
so this may let both work together. In any case, if the signature is not found
then the fallback method is selected, and is sufficient

Fixes #12